rename lib to utils
